crml_lang/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crml-lang"
version = "1.1.0"
description = "CRML language/spec models, schema validator, and YAML IO"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Zeron Research Labs"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "pyyaml",
  "jsonschema",
  "pydantic",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
crml_lang = ["schema/*.json"]
